Update parsers to adapt to dropped tool columns
