+++ parser/src/main/java/io/moderne/jenkins/ingest/Parser.java
@@ -47,7 +47,7 @@ public class Parser {
             // Loop over repoLines and update them based matching data table rows
             for (String line : reposLines) {
                 String[] split = line.split(",", -1);
-                CsvRow csvRow = new CsvRow(split[0], split[1], split[2], split[3], split[4], split[5], split[6], split[7], split[8], split[9]);
+                CsvRow csvRow = new CsvRow(split[0], split[1], split[2], split[3], split[4], split[5], split[6]);
                 DataTableRow datatableRow = datatableRowsByKey.get(new Key(csvRow.scmHost(), csvRow.repoName(), csvRow.repoBranch()));
                 if (datatableRow != null) {
                     csvRow = updateCsvRow(csvRow, datatableRow);
@@ -58,17 +58,6 @@ public class Parser {
     }
 
     private static CsvRow updateCsvRow(CsvRow csvRow, DataTableRow datatableRow) {
-        // Pick minimum required Java version
-        Integer requiredJavaVersion = javaVersion(datatableRow.requiredJavaVersion());
-        if (requiredJavaVersion != null) {
-            csvRow = csvRow.withJdkTool("java" + requiredJavaVersion);
-            if (requiredJavaVersion < 8) {
-                csvRow = csvRow
-                        .withRepoSkip("TRUE")
-                        .withSkipReason("Java version " + requiredJavaVersion + " is not supported");
-            }
-        }
-
         // Skip outdated Maven and Gradle wrappers
         String wrapper = datatableRow.type();
         String wrapperVersion = datatableRow.version();
@@ -86,44 +75,11 @@ public class Parser {
         return csvRow;
     }
 
-    private static Integer javaVersion(String requiredJavaVersion) {
-        if (requiredJavaVersion == null || requiredJavaVersion.isBlank()) {
-            return null;
-        }
-        int javaVersion = Integer.parseInt(requiredJavaVersion);
-        if (javaVersion <= 8) {
-            return javaVersion;
-        } else if (javaVersion <= 11) {
-            return 11;
-        } else if (javaVersion <= 17) {
-            return 17;
-        } else {
-            return 20;
-        }
+    record DataTableRow(String repositoryOrigin, String repositoryPath, String repositoryBranch, String type,
+                        String version, String command, String exitCode, String requiredJavaVersion, String logOutput) {
     }
 }
 
-record Key(String origin, String path, String branch) implements Comparable<Key> {
-    @Override
-    public int compareTo(Key o) {
-        return Comparator.comparing(Key::origin)
-                .thenComparing(Key::path)
-                .thenComparing(Key::branch)
-                .compare(this, o);
-    }
-}
 
-record DataTableRow(String repositoryOrigin, String repositoryPath, String repositoryBranch, String type,
-                    String version, String command, String exitCode, String requiredJavaVersion, String logOutput) {
-}
-
-@lombok.With
-record CsvRow(String scmHost, String repoName, String repoBranch, String mavenTool, String gradleTool, String jdkTool,
-              String repoStyle, String repoBuildAction, String repoSkip, String skipReason) {
-    @Override
-    public String toString() {
-        return String.join(",", scmHost, repoName, repoBranch, mavenTool, gradleTool, jdkTool, repoStyle, repoBuildAction, repoSkip, skipReason);
-    }
-}
